Simplify xterm.js (more code slice)
