+++ src/utils/off.js
@@ -1,0 +1,7 @@
+"use strict";
+
+function off(el, type, handler, capture) {
+  el.removeEventListener(type, handler, capture || false);
+}
+
+module.exports = off;
